requirements.txt
altgraph==0.17.4
ffmpeg-python==0.2.0
future==1.0.0
packaging==25.0
pydub==0.25.1
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.4
python-dateutil==2.9.0.post0
setuptools==80.3.1
six==1.17.0
